Moved heavy lifting classes to reditools folder instead of tool folders.
