+++ reditools/annotate/__init__.py
@@ -1,4 +1,3 @@
 from .main import main
-from .rtannotater import RTAnnotater
 
-__all__ = ('main', 'RTAnnotater')
+__all__ = ('main',)
