+++ reditools/index/__init__.py
@@ -1,4 +1,3 @@
 from .main import main
-from .rtindexer import RTIndexer
 
-__all__ = ('main', 'RTIndexer')
+__all__ = ('main',)
